requirements.txt
pandas
matplotlib
opencv-python
tensorflow
flask
sqlalchemy
gunicorn
pyyaml